updated seperate_space, fixed is_number, added coordinate to grid function
updated seperate_space function to check for numbers and if not a
number to return -1 and stderr error log.

fixed is_number as it was doing this weird thing where it wasn’t
working with strstr, so I changed it to strspn, don’t really know why
starts stopped working but the main thing is i figured out what wasn’t
working.

I created a add_coordinate_to_grid function. It adds the coordinates
from the play_game seperate_space returned values to the grid and then
prints the new grid with the current players symbol.

have also created a change_player function where it changes the current
player, so if current player is ‘X’ it is then changed to ‘O’.

Still got a lot of work to do with regards to the exploding moves and
checking for end of game.